requirements.txt
pandas==2.3.3
python-dateutil==2.9.0
rich==14.2.0
tabulate==0.9.0
python-dotenv==1.2.1
mistralai[agents]==1.10.0